feat: multi-environment setup (dev/uat/prod) + security hardening - Add 3 Spring profiles: dev, uat, prod with proper config isolation - Add IntelliJ run configurations: BillSmart [Dev], [UAT], [Prod] - Add Gradle tasks: runDev, runUat, runProd - Add JwtAuthenticationFilter and AppProperties config classes - Add application-uat.properties for UAT environment Security fixes: - Remove hardcoded DB password from application-dev.properties (now env var) - Remove hardcoded DB password from application-test.properties (now env var) - Move V17 seed data from db/migration to db/seed (only loaded in dev/uat via Flyway)   Prevents test users (admin123/staff123) from being seeded into production - DataInitializer restricted to @Profile(dev, uat) - will not run in prod - Prod profile uses only env vars for DB credentials and JWT secret Other: - Restore complete .gitignore (Spring IDE, VS Code, keystore exception) - Update .env.example with dev/uat/prod- Add 3 Spring profiles: dev, uat, prod with proper conq cd /Users/dilip/Documents/Projects/cloned/BillSmart && git commit -m 'feat: multi-environment setup (dev/uat/prod) + security hardening' printf '\003' exit cd /Users/dilip/Documents/Projects/cloned/BillSmart && git status

+++ src/main/resources/db/migration/V11__Create_order_items_table.sql
@@ -3,11 +3,11 @@
 -- Purpose: Individual items within an order (matches OrderItem entity exactly)
 
 CREATE TABLE IF NOT EXISTS order_items (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
 
     -- References
-    order_id INTEGER NOT NULL,
-    food_id INTEGER NOT NULL,
+    order_id BIGINT NOT NULL,
+    food_id BIGINT NOT NULL,
 
     -- Quantity & Pricing
     quantity INTEGER NOT NULL DEFAULT 1,
